# server-main/openapi.yaml
openapi: 3.0.3
info:
  title: Gestor de Tickets - Servidor Principal
  version: 1.0.0
  description: API REST para gestão de tickets e registo de webhooks.
servers:
  - url: http://localhost:3000
paths:
...
  /api/webhooks:
    post:
      summary: Registar webhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreateRequest'
      responses:
        '201':
          description: Webhook registado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookCreateResponse'
        '400':
          description: Pedido inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Listar webhooks registados
      tags: [Webhooks]
      responses:
        '200':
          description: Lista de webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookListItem'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      additionalProperties: false
...
    WebhookCreateRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        event:
          type: string
          enum: [ticket.created, ticket.updated, ticket.archived]
      required: [url, event]
      additionalProperties: false
    WebhookCreateResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
          description: 0 quando o registo é ignorado por duplicado
        url:
          type: string
          format: uri
        event:
          type: string
      required: [ok, id, url, event]
      additionalProperties: false
    WebhookListItem:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
        event:
          type: string
        active:
          type: integer
          enum: [0, 1]
        createdAt:
          type: string
          format: date-time
      required: [id, url, event, active, createdAt]
      additionalProperties: false